add callbacks to schedule directive
scheudle directive can now call a click and refresh handler from the
parent scope.
The refresh() callback can be used by a function getting data from a db
and updating the correct cell with a scheduled event, refresh() gets the
tbody element as parameter.
click() is called when user clicks on a cell, it gets the current day as
parameter.