db: new models and relations
Added an Upload model and renamed StudentAssignment to Submission and
normalized both to use the Upload model (which represents a file upload).
Also remade Schedule model to not have FK on type, subject and assignment,
instead use join tables.
ScheduleType was removed as a db model and made into a static enum.